weights as floats

+++ src/LatticeBoltzmann.hpp
@@ -27,7 +27,7 @@ private:
     Array2D<FICell> fi_Eq;
 
     // quadrature weights for approximating equilibrium distribution
-    int w [9] = {4/9, 1/9, 1/9, 1/9, 1/9, 1/36, 1/36, 1/36, 1/36};
+    float w [9] = {4./9., 1./9., 1./9., 1./9., 1./9., 1./36., 1./36., 1./36., 1./36.};
     int cx [9] = {0, 1, 0, -1, 0 , 1, -1, -1, 1};
     int cy [9] = {0, 0, 1, 0, -1, 1, 1, -1, -1};
     int opp [9] = {0, 3, 4, 1, 2, 7, 8, 5, 6};
